# internal/deviceops/transport/rest/zz_yunka_device_management_operation_executor_gen.go
// Code generated by yunka contract; DO NOT EDIT.

package rest

import (
	"errors"
	deviceopsv1 "github.com/hvritual/biz/contracts/gen/deviceops/v1"
	application "github.com/hvritual/biz/internal/deviceops/application"
	policy "github.com/hvritual/biz/internal/deviceops/policy"
	codes "google.golang.org/grpc/codes"
	status "google.golang.org/grpc/status"
	protojson "google.golang.org/protobuf/encoding/protojson"
	io "io"
	"net/http"
	strconv "strconv"
	execution "yunka.io/framework/execution"
	operation "yunka.io/framework/operation"
	authz "yunka.io/gateway/authz"
	httpbinding "yunka.io/gateway/httpbinding"
)

type DeviceManagementOperationHandler struct {
	application application.DeviceManagementApplication
	executor    operation.Executor
}

func RegisterDeviceManagementOperationExecutor(mux *http.ServeMux, application application.DeviceManagementApplication, executor operation.Executor) error {
	if mux == nil {
		return errors.New("contract C9 REST adapter: mux is required")
	}
	if application == nil {
		return errors.New("contract C9 REST adapter: application is required")
	}
	if executor == nil {
		return errors.New("contract C9 REST adapter: operation executor is required")
	}
	handler := &DeviceManagementOperationHandler{application: application, executor: executor}
	if err := httpbinding.Register(mux, "POST", "/v1/devices", handler.handleOperationCreateDevice); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "DELETE", "/v1/devices/{id}", handler.handleOperationDeleteDevice); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "GET", "/v1/devices/{id}", handler.handleOperationGetDevice); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "GET", "/v1/devices", handler.handleOperationListDevices); err != nil {
		return err
	}
	if err := httpbinding.Register(mux, "PATCH", "/v1/devices/{id}", handler.handleOperationUpdateDevice); err != nil {
		return err
	}
	return nil
}

func writeDeviceManagementOperationError(writer http.ResponseWriter, err error) {
	if authz.IsDenied(err) {
		statusCode := http.StatusForbidden
		var denied *authz.DeniedError
		if errors.As(err, &denied) && (denied.Decision.Reason == authz.ReasonUnauthenticated || denied.Decision.Reason == authz.ReasonAuthenticationMethod) {
			statusCode = http.StatusUnauthorized
		}
		http.Error(writer, http.StatusText(statusCode), statusCode)
		return
	}
	if errors.Is(err, execution.ErrIdempotencyKeyRequired) {
		http.Error(writer, "idempotency key required", http.StatusBadRequest)
		return
	}
	if errors.Is(err, execution.ErrIdempotencyInProgress) || errors.Is(err, execution.ErrIdempotencyCompleted) {
		http.Error(writer, "idempotency conflict", http.StatusConflict)
		return
	}
	if code := status.Code(err); code == codes.Aborted || code == codes.AlreadyExists {
		http.Error(writer, "application conflict", http.StatusConflict)
		return
	}
	if errors.Is(err, operation.ErrExecutorUnavailable) || errors.Is(err, operation.ErrSecurityUnavailable) || errors.Is(err, operation.ErrSecurityNilContext) || errors.Is(err, operation.ErrIdempotencyUnavailable) {
		http.Error(writer, "operation execution unavailable", http.StatusInternalServerError)
		return
	}
	http.Error(writer, "application request failed", http.StatusBadRequest)
}

func (handler *DeviceManagementOperationHandler) handleOperationCreateDevice(writer http.ResponseWriter, request *http.Request) {
	wire := &deviceopsv1.CreateDeviceRequest{}
	body, err := io.ReadAll(request.Body)
	if err != nil {
		http.Error(writer, "invalid request body", http.StatusBadRequest)
		return
	}
	if len(body) > 0 {
		if err := protojson.Unmarshal(body, wire); err != nil {
			http.Error(writer, "invalid request body", http.StatusBadRequest)
			return
		}
	}
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanDeviceManagementCreateDevice(), wire, handler.application.CreateDevice)
	if err != nil {
		writeDeviceManagementOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *DeviceManagementOperationHandler) handleOperationDeleteDevice(writer http.ResponseWriter, request *http.Request) {
	wire := &deviceopsv1.DeleteDeviceRequest{}
	if raw := request.URL.Query().Get("version"); raw != "" {
		parsed, err := strconv.ParseUint(raw, 10, 64)
		if err != nil {
			http.Error(writer, "invalid request parameter", http.StatusBadRequest)
			return
		}
		wire.Version = uint64(parsed)
	}
	wire.Id = request.PathValue("id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanDeviceManagementDeleteDevice(), wire, handler.application.DeleteDevice)
	if err != nil {
		writeDeviceManagementOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *DeviceManagementOperationHandler) handleOperationGetDevice(writer http.ResponseWriter, request *http.Request) {
	wire := &deviceopsv1.GetDeviceRequest{}
	wire.Id = request.PathValue("id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanDeviceManagementGetDevice(), wire, handler.application.GetDevice)
	if err != nil {
		writeDeviceManagementOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *DeviceManagementOperationHandler) handleOperationListDevices(writer http.ResponseWriter, request *http.Request) {
	wire := &deviceopsv1.ListDevicesRequest{}
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanDeviceManagementListDevices(), wire, handler.application.ListDevices)
	if err != nil {
		writeDeviceManagementOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}

func (handler *DeviceManagementOperationHandler) handleOperationUpdateDevice(writer http.ResponseWriter, request *http.Request) {
	wire := &deviceopsv1.UpdateDeviceRequest{}
	body, err := io.ReadAll(request.Body)
	if err != nil {
		http.Error(writer, "invalid request body", http.StatusBadRequest)
		return
	}
	if len(body) > 0 {
		if err := protojson.Unmarshal(body, wire); err != nil {
			http.Error(writer, "invalid request body", http.StatusBadRequest)
			return
		}
	}
	wire.Id = request.PathValue("id")
	callContext := execution.WithIdempotencyKey(request.Context(), request.Header.Get("Idempotency-Key"))
	output, err := operation.ExecuteTyped(callContext, handler.executor, policy.OperationPlanDeviceManagementUpdateDevice(), wire, handler.application.UpdateDevice)
	if err != nil {
		writeDeviceManagementOperationError(writer, err)
		return
	}
	payload, err := protojson.Marshal(output)
	if err != nil {
		http.Error(writer, "response encoding failed", http.StatusInternalServerError)
		return
	}
	writer.Header().Set("Content-Type", "application/json")
	_, _ = writer.Write(payload)
}